Change mysqlfs setup and service startup to honor MySQL port
To maintain backward compatibility, we need to pass 3306 as the port when the
MYSQLFS_PORT variable is not set. ExecStart directive cannot evaluate bash
directives. The service startup complains because of that syntax. As a
workaround, we pass the port explicitly as an environment variable and then
override it through the environment file.